requirements.txt
streamlit
pytesseract
pdf2image
pandas
xlsxwriter
pillow